requirements.txt
pillow==11.1.0
tifffile==2024.8.30
memray
pytest
pypng==0.20220715.0
